llm/requirements.txt
pinecone-client==3.0.0
sentence-transformers==2.6.1
torch==2.2.1
transformers==4.38.2
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.66.2
langgraph==0.0.30

